Abstract outL-iso and fill some of the holes
As pointed out by Régis, agda is choking when normalising inL.
Abstracting the definition of outL-iso makes most of the definitions
involving inL and outL typecheck in reasonable time.

Of course, the problem is now proving the lemmas that actually use the
definition of outL-iso, like Φ-Ψ-comm.  However, this should be doable
by equational reasoning one we abstract out and prove basic commutation
properties of this isomorphism.